Fix TypeScript typing error
This was giving me an error in my IDE. I'm not sure if this is the correct solution though, but it silences the error. The error may lie with the CredentialsProvider class and simply fixing the typing to allow a number instead of a string, but that is an upstream dependency.